Safari-611.1.10/LayoutTests/webaudio/oscillatornode-legacy-api-expected.txt

Testing that the legacy OscillatorNode API is not exposed on the unprefixed OscillatorNode

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS OscillatorNode.prototype.playbackState is undefined.
PASS OscillatorNode.UNSCHEDULED_STATE is undefined.
PASS OscillatorNode.SCHEDULED_STATE is undefined.
PASS OscillatorNode.PLAYING_STATE is undefined.
PASS OscillatorNode.FINISHED_STATE is undefined.
PASS OscillatorNode.prototype.UNSCHEDULED_STATE is undefined.
PASS OscillatorNode.prototype.SCHEDULED_STATE is undefined.
PASS OscillatorNode.prototype.PLAYING_STATE is undefined.
PASS OscillatorNode.prototype.FINISHED_STATE is undefined.
PASS oscillator.playbackState is undefined.
PASS oscillator.UNSCHEDULED_STATE is undefined.
PASS oscillator.SCHEDULED_STATE is undefined.
PASS oscillator.PLAYING_STATE is undefined.
PASS oscillator.FINISHED_STATE is undefined.
PASS oscillator.__proto__.playbackState is undefined.
PASS oscillator.__proto__.UNSCHEDULED_STATE is undefined.
PASS oscillator.__proto__.SCHEDULED_STATE is undefined.
PASS oscillator.__proto__.PLAYING_STATE is undefined.
PASS oscillator.__proto__.FINISHED_STATE is undefined.
PASS successfullyParsed is true

TEST COMPLETE
